+++ kjar/pom.xml
@@ -1,11 +1,11 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
    <modelVersion>4.0.0</modelVersion>
    <parent>
-      <groupId>io.silverware.demos</groupId>
-      <artifactId>gateway</artifactId>
-      <version>2.0-SNAPSHOT</version>
+      <groupId>com.redhat.patriot</groupId>
+      <artifactId>smart-home-gateway</artifactId>
+      <version>1.0-SNAPSHOT</version>
    </parent>
-   <artifactId>gateway-kjar</artifactId>
+   <artifactId>smart-home-gateway-kjar</artifactId>
    <packaging>kjar</packaging>
    <dependencies>
       <dependency>
