+++ kjar/src/main/java/com/redhat/patriot/smart_home_gateway/kjar/FireplaceCommand.java
@@ -1,0 +1,72 @@
+/*
+ * Copyright 2018 Patriot project
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+package com.redhat.patriot.smart_home_gateway.kjar;
+
+import java.util.Collections;
+import java.util.Map;
+
+/**
+ * @author <a href="mailto:marvenec@gmail.com">Martin Večeřa</a>
+ */
+public class FireplaceCommand extends Command {
+
+   public enum Fire {
+      HEAT, COLD
+   }
+
+   private Fire fire = Fire.COLD;
+
+   public FireplaceCommand(final Fire fire) {
+      this.fire = fire;
+   }
+
+   public Fire getFire() {
+      return fire;
+   }
+
+   @Override
+   public boolean equals(final Object o) {
+      if (this == o) {
+         return true;
+      }
+      if (o == null || getClass() != o.getClass()) {
+         return false;
+      }
+
+      final FireplaceCommand that = (FireplaceCommand) o;
+
+      return fire == that.fire;
+
+   }
+
+   @Override
+   public int hashCode() {
+      return fire.hashCode();
+   }
+
+   @Override
+   public String toString() {
+      return "FireplaceCommand{" +
+            "fire=" + fire +
+            '}';
+   }
+
+   @Override
+   public Map<String, String> getCacheUpdate() {
+      return Collections.singletonMap(this.getClass().getCanonicalName(), fire.toString());
+   }
+
+}
