+++ kjar/src/main/java/com/redhat/patriot/smart_home_gateway/kjar/Weather.java
@@ -1,0 +1,74 @@
+/*
+ * Copyright 2018 Patriot project
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+package com.redhat.patriot.smart_home_gateway.kjar;
+
+import java.io.Serializable;
+
+/**
+ * @author <a href="mailto:marvenec@gmail.com">Martin Večeřa</a>
+ */
+public class Weather implements Serializable {
+
+   private int temperature;
+   private int humidity;
+
+   public Weather(final int temperature, final int humidity) {
+      this.temperature = temperature;
+      this.humidity = humidity;
+   }
+
+   public int getTemperature() {
+      return temperature;
+   }
+
+   public int getHumidity() {
+      return humidity;
+   }
+
+   @Override
+   public boolean equals(final Object o) {
+      if (this == o) {
+         return true;
+      }
+      if (o == null || getClass() != o.getClass()) {
+         return false;
+      }
+
+      final Weather that = (Weather) o;
+
+      if (temperature != that.temperature) {
+         return false;
+      }
+      return humidity == that.humidity;
+
+   }
+
+   @Override
+   public int hashCode() {
+      int result = temperature;
+      result = 31 * result + humidity;
+      return result;
+   }
+
+   @Override
+   public String toString() {
+      return "Weather{" +
+            "temperature=" + temperature +
+            ", humidity=" + humidity +
+            '}';
+   }
+
+}
