--- kjar/src/main/java/io/silverware/demos/devconf/kjar/CacheStatus.java
@@ -1,75 +1,0 @@
-/*
- * -----------------------------------------------------------------------\
- * SilverWare
- *  
- * Copyright (C) 2014 - 2016 the original author or authors.
- *  
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * -----------------------------------------------------------------------/
- */
-package io.silverware.demos.devconf.kjar;
-
-/**
- * @author <a href="mailto:marvenec@gmail.com">Martin Večeřa</a>
- */
-public class CacheStatus {
-
-   private final String key;
-   private final String value;
-
-   public CacheStatus(final String key, final String value) {
-      this.key = key;
-      this.value = value;
-   }
-
-   public String getKey() {
-      return key;
-   }
-
-   public String getValue() {
-      return value;
-   }
-
-   @Override
-   public boolean equals(final Object o) {
-      if (this == o) {
-         return true;
-      }
-      if (o == null || getClass() != o.getClass()) {
-         return false;
-      }
-
-      final CacheStatus that = (CacheStatus) o;
-
-      if (key != null ? !key.equals(that.key) : that.key != null) {
-         return false;
-      }
-      return value != null ? value.equals(that.value) : that.value == null;
-
-   }
-
-   @Override
-   public int hashCode() {
-      int result = key != null ? key.hashCode() : 0;
-      result = 31 * result + (value != null ? value.hashCode() : 0);
-      return result;
-   }
-
-   @Override
-   public String toString() {
-      return "CacheStatus{" +
-            "key='" + key + '\'' +
-            ", value='" + value + '\'' +
-            '}';
-   }
-}
