--- kjar/src/main/java/io/silverware/demos/devconf/kjar/UpdateStatusCommand.java
@@ -1,90 +1,0 @@
-/*
- * -----------------------------------------------------------------------\
- * SilverWare
- *  
- * Copyright (C) 2010 - 2013 the original author or authors.
- *  
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * -----------------------------------------------------------------------/
- */
-package io.silverware.demos.devconf.kjar;
-
-/**
- * @author <a href="mailto:marvenec@gmail.com">Martin Večeřa</a>
- */
-public class UpdateStatusCommand extends Command {
-
-   private String itemId = "";
-   private long color = 0xeeeeee;
-   private String status = "";
-
-   public UpdateStatusCommand(final String itemId, final long color, final String status) {
-      this.itemId = itemId;
-      this.color = color;
-      this.status = status;
-   }
-
-   public String getItemId() {
-      return itemId;
-   }
-
-   public long getColor() {
-      return color;
-   }
-
-   public String getStatus() {
-      return status;
-   }
-
-   public String getUpdateMessage() {
-      return itemId + ":" + color + ":" + status;
-   }
-
-   @Override
-   public boolean equals(final Object o) {
-      if (this == o) {
-         return true;
-      }
-      if (o == null || getClass() != o.getClass()) {
-         return false;
-      }
-
-      final UpdateStatusCommand that = (UpdateStatusCommand) o;
-
-      if (color != that.color) {
-         return false;
-      }
-      if (!itemId.equals(that.itemId)) {
-         return false;
-      }
-      return status.equals(that.status);
-
-   }
-
-   @Override
-   public int hashCode() {
-      int result = itemId.hashCode();
-      result = 31 * result + (int) (color ^ (color >>> 32));
-      result = 31 * result + status.hashCode();
-      return result;
-   }
-
-   @Override
-   public String toString() {
-      return "UpdateStatusCommand{" +
-            "itemId='" + itemId + '\'' +
-            ", color=" + color +
-            ", status='" + status + '\'' +
-            '}';
-   }
-}
